+++ src/main/java/nl/delftelectronics/spaceinvaders/core/scenes/GameScene.java
@@ -129,8 +129,8 @@ public class GameScene implements EntityDestroyedListener {
 	public List<Entity> getCollisions(Collidable collidee) {
 		ArrayList<Entity> result = new ArrayList<Entity>();
 
-		Rectangle referenceRectangle = new Rectangle(collidee.getPositionX(), collidee.getPositionY(),
-				collidee.getWidth(), collidee.getHeight());
+		Rectangle referenceRectangle = new Rectangle((int)collidee.getPositionX(), (int)collidee.getPositionY(),
+				(int)collidee.getWidth(), (int)collidee.getHeight());
 
 		for (Entity candidate : entities) {
 			if (candidate == collidee) {
@@ -138,8 +138,8 @@ public class GameScene implements EntityDestroyedListener {
 			}
 			if (candidate instanceof Collidable) {
 				Collidable c = (Collidable) candidate;
-				Rectangle testRectangle = new Rectangle(c.getPositionX(), c.getPositionY(), c.getWidth(),
-						c.getHeight());
+				Rectangle testRectangle = new Rectangle((int)c.getPositionX(), (int)c.getPositionY(),
+						(int)c.getWidth(), (int)c.getHeight());
 
 				if (referenceRectangle.intersects(testRectangle)) {
 					result.add(candidate);
